send email for consultants and contacts were registered

+++ src/main/java/co/edu/ufps/innova/consultant/domain/service/ConsultantService.java
@@ -2,10 +2,13 @@ package co.edu.ufps.innova.consultant.domain.service;
 
 import java.util.List;
 import java.util.Optional;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import co.edu.ufps.innova.email.domain.dto.Email;
 import co.edu.ufps.innova.user.domain.service.UserService;
 import co.edu.ufps.innova.consultant.domain.dto.Consultant;
+import co.edu.ufps.innova.email.domain.repository.IEmailRepository;
 import co.edu.ufps.innova.consultant.domain.repository.IConsultantRepository;
 import co.edu.ufps.innova.authentication.domain.repository.IPasswordRepository;
 
@@ -15,17 +18,23 @@ public class ConsultantService {
 
     private final UserService userService;
     private final IConsultantRepository repository;
+    private final IEmailRepository IEmailRepository;
     private final IPasswordRepository passwordRepository;
 
     public Consultant save(Consultant consultant) {
-        String userPassword;
-        if (userService.findById(consultant.getId()).isPresent()) {
-            userPassword = userService.getPassword(consultant.getId());
-            userService.delete(consultant.getId());
-        } else {
-            userPassword = passwordRepository.encryptPassword(passwordRepository.generatePassword());
-        }
-        return repository.save(consultant, userPassword);
+        if (userService.findById(consultant.getId()).isPresent()) userService.delete(consultant.getId());
+        String userPassword = passwordRepository.generatePassword();
+        Consultant myConsultant = repository.save(consultant, passwordRepository.encryptPassword(userPassword));
+
+        Email email = new Email();
+        email.setTo(myConsultant.getEmail());
+        email.setSubject("Innova - Registro exitoso");
+        email.setContent("Con esta contraseña podrá ingresar a la plataforma: \n " +
+                "\t " + userPassword +
+                "\n recomendamos cambie esta contraseña desde la plataforma apenas ingrese a la misma.");
+        IEmailRepository.sendEmail(email);
+
+        return myConsultant;
     }
 
     public boolean update(String id, Consultant consultant) {
