+++ src/main/java/co/edu/ufps/innova/contact/domain/service/ContactService.java
@@ -3,11 +3,14 @@ package co.edu.ufps.innova.contact.domain.service;
 import java.util.List;
 import java.util.Optional;
 import java.time.LocalDate;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import co.edu.ufps.innova.email.domain.dto.Email;
 import co.edu.ufps.innova.contact.domain.dto.Contact;
 import co.edu.ufps.innova.contact.domain.dto.ContactType;
 import co.edu.ufps.innova.user.domain.service.UserService;
+import co.edu.ufps.innova.email.domain.repository.IEmailRepository;
 import co.edu.ufps.innova.contact.domain.repository.IContactRepository;
 import co.edu.ufps.innova.authentication.domain.repository.IPasswordRepository;
 
@@ -17,18 +20,24 @@ public class ContactService {
 
     private final UserService userService;
     private final IContactRepository repository;
+    private final IEmailRepository IEmailRepository;
     private final IPasswordRepository passwordRepository;
 
     public Contact save(Contact contact) {
-        String userPassword;
-        if (userService.findById(contact.getId()).isPresent()) {
-            userPassword = userService.getPassword(contact.getId());
-            userService.delete(contact.getId());
-        } else {
-            userPassword = passwordRepository.encryptPassword(passwordRepository.generatePassword());
-        }
+        if (userService.findById(contact.getId()).isPresent()) userService.delete(contact.getId());
         contact.setRegistrationDate(LocalDate.now());
-        return repository.save(contact, userPassword);
+        String userPassword = passwordRepository.generatePassword();
+        Contact myContact = repository.save(contact, passwordRepository.encryptPassword(userPassword));
+
+        Email email = new Email();
+        email.setTo(myContact.getEmail());
+        email.setSubject("Innova - Registro exitoso");
+        email.setContent("Con esta contraseña podrá ingresar a la plataforma: \n " +
+                "\t " + userPassword +
+                "\n recomendamos cambie esta contraseña desde la plataforma apenas ingrese a la misma.");
+        IEmailRepository.sendEmail(email);
+
+        return myContact;
     }
 
     public boolean update(String id, Contact contact) {
